tmfcore_war/pom.xml: fix jersey and jetty
It was wrong to declare jetty as a compile dependency because
we're using it as an external container (which could be replaced
for example by tomcat).

Change the jersey version to 2.7. This is arbitrary.

Also, do not use scope 'import' which was not the correct
thing to do in this context.