Split @thermz's HTTP code into multiple files
Hooray hooray, the current implementation deploys and works
on Tomcat 8 using Java 8 on Xubuntu 14.10.

NOTE: the critical piece of code that was lacking to @thermz
code to be run on Tomcat is the implementation of getClasses()
method of the class overriding Application. I have seen this method
reimplemented by many examples and tutorials and indeed this was
the piece required in order ot make it work on Tomcat.

While there, add a copyright header and give full credit to
@thermz; my contribution here is small.

Also .gitignore generated files.

I also needed to include Jersey's binding to Jackson so
that we can read/write JSON.